test(zqlite): customize education app scenario seeds
