+++ packages/zqlite/src/test/query-scenarios/scenarios/or-with-false-branch.ts
@@ -1,10 +1,21 @@
 import type {QueryScenario} from '../../query-scenario.ts';
-import {assignmentSchema, seedAssignments} from '../assignment.ts';
+import {
+  educationAppSchema,
+  createEducationAppTables,
+} from '../education-app.ts';
 
 export default {
   name: 'OR with false branch keeps surviving teacher predicate',
-  schema: assignmentSchema,
-  seed: seedAssignments,
+  schema: educationAppSchema,
+  seed: db => {
+    createEducationAppTables(db);
+
+    const assignmentStmt = db.prepare(
+      'INSERT INTO assignment (id, teacher_id, archived_at, created_at) VALUES (?, ?, ?, ?)',
+    );
+    assignmentStmt.run(1, 1, null, 1);
+    assignmentStmt.run(2, 2, null, 2);
+  },
   query: builder =>
     builder.assignment
       .where(({cmp, or}) => or(cmp('teacher_id', '=', 1), or()))
@@ -26,4 +37,4 @@ export default {
       },
     ],
   },
-} satisfies QueryScenario<typeof assignmentSchema>;
+} satisfies QueryScenario<typeof educationAppSchema>;
